requirements.txt
tensorflow-gpu==1.2.1
tensorflow-tensorboard==0.1.7
imageio==2.2.0
